Removed interface TargetManager
