+++ webcurator-core/src/main/java/org/webcurator/core/coordinator/WctCoordinator.java
@@ -42,6 +42,7 @@ import org.webcurator.core.notification.InTrayManager;
 import org.webcurator.core.scheduler.TargetInstanceManager;
 import org.webcurator.core.store.DigitalAssetStoreFactory;
 import org.webcurator.core.store.coordinator.DigitalAssetStoreCoordinator;
+import org.webcurator.core.targets.TargetManager;
 import org.webcurator.core.util.AuthUtil;
 import org.webcurator.core.util.PatchUtil;
 import org.webcurator.core.util.WctUtils;
@@ -53,7 +54,6 @@ import org.webcurator.core.notification.MessageType;
 import org.webcurator.core.profiles.Heritrix3Profile;
 import org.webcurator.core.profiles.HeritrixProfile;
 import org.webcurator.core.store.DigitalAssetStore;
-import org.webcurator.core.targets.TargetManager;
 import org.webcurator.core.visualization.VisualizationConstants;
 import org.webcurator.core.visualization.networkmap.metadata.NetworkMapResult;
 import org.webcurator.core.visualization.networkmap.service.NetworkMapClient;
